Types changed!

+++ 01.C# Part 1/03.Operators-and-Expressions/14.ModifyABitAtGivenPosition/ModifyABitAtGivenPosition.cs
@@ -4,20 +4,20 @@ class ModifyABit
 {
     static void Main()
     {
-        int number = int.Parse(Console.ReadLine());
-        int position = int.Parse(Console.ReadLine());
-        int value = int.Parse(Console.ReadLine());
+        ulong number = ulong.Parse(Console.ReadLine());
+        byte position = byte.Parse(Console.ReadLine());
+        byte value = byte.Parse(Console.ReadLine());
 
         if (value == 0)
         {
-            int mask = ~(1 << position);
-            int result = number & mask;
+            ulong mask = (ulong)~(1 << position);
+            ulong result = number & mask;
             Console.WriteLine(result);
         }
         else
         {
-            int mask = 1 << position;
-            int result = number | mask;
+            ulong mask = (ulong)1 << position;
+            ulong result = number | mask;
             Console.WriteLine(result);
         }
     }
